leaning fully on TSC to build .d.ts files
