+++ packages/integrations/image/components/Image.astro
@@ -1,13 +1,13 @@
 ---
 // @ts-ignore
-import { getImage } from '../dist/index.js';
-import type { ImageAttributes, ImageMetadata, TransformOptions, OutputFormat } from '../dist/types';
+import { getImage } from '../src/get-image.js';
+import type { ImageMetadata, TransformOptions, OutputFormat } from '../src/index.js';
 
-export interface LocalImageProps extends Omit<TransformOptions, 'src'>, Omit<ImageAttributes, 'src' | 'width' | 'height'> {
+export interface LocalImageProps extends Omit<TransformOptions, 'src'>, Omit<astroHTML.JSX.ImgHTMLAttributes, 'src' | 'width' | 'height'> {
 	src: ImageMetadata | Promise<{ default: ImageMetadata }>;
 }
 
-export interface RemoteImageProps extends TransformOptions, ImageAttributes {
+export interface RemoteImageProps extends TransformOptions, astroHTML.JSX.ImgHTMLAttributes {
 	src: string;
 	format: OutputFormat;
 	width: number;
