+++ packages/integrations/image/components/Picture.astro
@@ -1,8 +1,11 @@
 ---
-import { getPicture } from '../dist/index.js';
-import type { ImageAttributes, ImageMetadata, OutputFormat, PictureAttributes, TransformOptions } from '../dist/types';
+import { getPicture } from '../src/index.js';
+import type { ImageMetadata, OutputFormat, TransformOptions } from '../src/index.js';
 
-export interface LocalImageProps extends Omit<PictureAttributes, 'src' | 'width' | 'height'>, Omit<TransformOptions, 'src'>, Pick<ImageAttributes, 'loading' | 'decoding'> {
+export interface LocalImageProps extends
+	Omit<astroHTML.JSX.HTMLAttributes, 'src' | 'width' | 'height'>,
+	Omit<TransformOptions, 'src'>,
+	Pick<astroHTML.JSX.ImgHTMLAttributes, 'loading' | 'decoding'> {
 	src: ImageMetadata | Promise<{ default: ImageMetadata }>;
 	alt?: string;
 	sizes: HTMLImageElement['sizes'];
@@ -10,7 +13,10 @@ export interface LocalImageProps extends Omit<PictureAttributes, 'src' | 'width'
 	formats?: OutputFormat[];
 }
 
-export interface RemoteImageProps extends Omit<PictureAttributes, 'src' | 'width' | 'height'>, TransformOptions, Pick<ImageAttributes, 'loading' | 'decoding'> {
+export interface RemoteImageProps extends
+	Omit<astroHTML.JSX.HTMLAttributes, 'src' | 'width' | 'height'>,
+	TransformOptions,
+	Pick<astroHTML.JSX.ImgHTMLAttributes, 'loading' | 'decoding'> {
 	src: string;
 	alt?: string;
 	sizes: HTMLImageElement['sizes'];
