+++ packages/integrations/image/src/lib/get-image.ts
@@ -1,15 +1,15 @@
+/// <reference types="astro/astro-jsx" />
 import slash from 'slash';
 import { ROUTE_PATTERN } from '../constants.js';
 import sharp from '../loaders/sharp.js';
 import {
-	ImageAttributes,
-	ImageMetadata,
 	ImageService,
 	isSSRService,
 	OutputFormat,
 	TransformOptions,
-} from '../types.js';
+} from '../loaders/index.js';
 import { isRemoteImage, parseAspectRatio } from '../utils/images.js';
+import { ImageMetadata } from '../vite-plugin-astro-image.js';
 
 export interface GetImageTransform extends Omit<TransformOptions, 'src'> {
 	src: string | ImageMetadata | Promise<{ default: ImageMetadata }>;
@@ -101,7 +101,7 @@ async function resolveTransform(input: GetImageTransform): Promise<TransformOpti
  * @param transform @type {TransformOptions} The transformations requested for the optimized image.
  * @returns @type {ImageAttributes} The HTML attributes to be included on the built `<img />` element.
  */
-export async function getImage(transform: GetImageTransform): Promise<ImageAttributes> {
+export async function getImage(transform: GetImageTransform): Promise<astroHTML.JSX.ImgHTMLAttributes> {
 	if (!transform.src) {
 		throw new Error('[@astrojs/image] `src` is required');
 	}
