+++ packages/integrations/image/src/lib/get-picture.ts
@@ -1,7 +1,9 @@
+/// <reference types="astro/astro-jsx" />
 import { lookup } from 'mrmime';
 import { extname } from 'node:path';
-import { ImageAttributes, ImageMetadata, OutputFormat, TransformOptions } from '../types.js';
+import { OutputFormat, TransformOptions } from '../loaders/index.js';
 import { parseAspectRatio } from '../utils/images.js';
+import { ImageMetadata } from '../vite-plugin-astro-image.js';
 import { getImage } from './get-image.js';
 
 export interface GetPictureParams {
@@ -12,7 +14,7 @@ export interface GetPictureParams {
 }
 
 export interface GetPictureResult {
-	image: ImageAttributes;
+	image: astroHTML.JSX.HTMLAttributes;
 	sources: { type: string; srcset: string }[];
 }
