+++ packages/integrations/image/src/vite-plugin-astro-image.ts
@@ -3,8 +3,16 @@ import { pathToFileURL } from 'node:url';
 import type { PluginContext } from 'rollup';
 import slash from 'slash';
 import type { Plugin, ResolvedConfig } from 'vite';
-import type { IntegrationOptions } from './types.js';
+import type { IntegrationOptions } from './index.js';
+import type { InputFormat } from './loaders/index.js';
 import { metadata } from './utils/metadata.js';
+
+export interface ImageMetadata {
+	src: string;
+	width: number;
+	height: number;
+	format: InputFormat;
+}
 
 export function createPlugin(config: AstroConfig, options: Required<IntegrationOptions>): Plugin {
 	const filter = (id: string) =>
